Added component context external links to edit view
Also removed '?' from "Please indicate what services you expect to utilize as part of your project?" on the create page. Adding question marks to sentences that are declarative or imperative is a linguistic choice of cowards who desperately try to remove themselves of ownership of their words. It's basically like the phrase "Just sayin'".